Condense textcolo(u)r() into text()
The way it works now is that text() will turn into textcolo(u)r() if you
give it 2 extra arguments, so basically you no longer have to type
textcolo(u)r() and can just type text() if you want to use
textcolo(u)r() now. If you give text() 2 extra arguments, it will start
interpreting arguments 1 through 3 as the R, G, and B, instead of
interpreting argument 1 as the color name and 2 and 3 as the X and Y
positions.