Move game.customtrials to ed.customtrials
The vector of custom trials for the current level seems like it should
belong in the 'ed' object with the rest of the other
currently-loaded-custom-level stuff, but putting it on the 'game' object
is really awkward, because that means it doesn't get reset when 'ed'
does, and so there had to be a special line of code to reset
game.customtrials SPECIFICALLY in case 20 when going to the level editor
from the main menu.

But now that it's moved onto 'ed', it will get naturally reset when 'ed'
gets reset, meaning we cover all cases we might need to reset
customtrials when 'ed' gets reset.