Ignore initial reports on whether the OSK is up
SDL doesn't seem to have a way to know whether a hardware keyboard or an
on-screen keyboard is in use. A delay would possibly solve this, but
this method works as well and prevents race conditions. This has a
possibility of breaking hardware keyboards on Android, however.
If you have a better idea, please suggest it.