Add dimensions: read/write to level, implementation, and gotodimension()
A dimension is a rectangle of rooms within which the game will keep you
in, but will otherwise let you wrap around if you reach an edge. I'm
also planning on making it so the minimap rendered is only of the
current dimension.

This commit adds support for reading and writing them to the level file,
the actual functionality of being inside a dimension, and the script
command gotodimension().

To be added later is support for setting, modifying, and removing
dimensions in the level editor, and restricting the minimap to only draw
the rooms in the current dimension.

In the custom level XML, dimensions are stored in the <dimensions> tag.
Each dimension looks like the following:

    <dimension x="0" y="0" w="2" h="2">

This dimension is one that starts at 0,0 and is 2x2.

By default, you are in dimension -1, which is just an invalid dimension
and so the game will treat you like you aren't in a dimension. You can
change the dimension you are currently in by using the script command
gotodimension(). The number you give to it is the ID of the dimension.
Dimension IDs are 0-indexed, and are determined by their order in the
<dimensions> tag.

If you set the ID of the current dimension to be an invalid one, the
game will ignore it and will act like you are not in a dimension.

A valid dimension has to have positive width and positive height. In
addition, dimensions CAN wrap around the map, but they CANNOT overlap
themselves. So that means their width must be less than or equal to the
width of the map, and (separately) their height must be less than or
equal to the height of the map.

Once you are in a dimension, as long as it is a valid dimension, the
game will wrap you around every time a gotoroom happens, and the
targeted room is outside the current dimension.