Fix minimap rendering in dimensions
Now it will render only the rooms in the dimension and nothing else.

If you're outside the dimension but haven't moved into a room inside it
yet, your cursor on the map will be outside the minimap. You probably
shouldn't be outside the dimension anyways.

Note that this does NOT fix it for the teleporter map. You can still
access teleporters from everywhere on the map.