Reset script delay when loading a script if it's a pdelay
This essentially makes it so when you activate a script via script box
or terminal in the middle of a pdelay, it runs immediately. There's no
way that you can load a script with load() in the middle of a pdelay,
because it's a script delay, so there's no problems with doing this
otherwise.