Correct terminal activity zone misalignment in VCE levels
For some strange reason, in vanilla, terminals' activity zones are
off-center by 8 pixels to the left, and they have 4 more pixels of width
than necessary, resulting in a strange activity zone alignment with the
terminal entity.

To avoid breaking any level that could rely on this misalignment, fix it
only in VCE levels.