Clear callstack when current script stops running
The intention of the previous line here was to pop a stack frame when a
script finished, to add an implicit return() at the end of every script.
However, this didn't work properly, and instead what happened is that
the call stack slowly disappeared away over several frames instead of
immediately disappearing. Also, this implict return() breaks convention
with regards to script writing. So I'm simply replacing it with a
callstack.clear().