Fix double-calling oneline internal scripts twice
In vanilla, there's a neat trick with internal commands, where if you
have an internal script that's only one command, you don't have to
bother writing the internal script formatting, and you can just
iftrinkets() to it. This uses the same exploit responsible for
no-cutscene-bars loadscript internal scripting in the first place.

Here's how it worked in vanilla, using createcrewman() as an example:
the iftrinkets() would happen, then the line gets ran because of an
indiscriminant tokenize() plus lucky frame ordering (createcrewman() is
checked after customiftrinkets()). However, then the createcrewman() in
the actual script is discarded due to not being a simplified command.

What changed in VCE is that non-simplified commands are no longer
discarded (because VCE lets you use internal commands directly in
simplified), so first the loadscript exploits happens, then the actual
line of createcrewman() itself is ran, thus resulting in spawning one
crewmate twice. With the default graphics, this isn't visible, until the
crewmate moves and leaves behind the other duplicate crewmate.

So to fix this, only allow using internal commands in simplified if
we're in a VCE level.