Store playtesting ghosts per-room
Now, when you go through one room, the ghosts for that room will stay
there. Going between rooms, in either in-game or in the editor, will no
longer reset the ghosts (I made sure to reset the ghosts when starting
playtesting, though).

And this also fixes a bug where if you exited playtesting while in a
different room than where you started playtesting, the ghosts for that
room would still be overlaid on the screen, even though it's incorrect.