Fix loadscript frame-ordering bug for VCE commands
Here's how you would internally script without cutscene bars before VCE:
you had to use iftrinkets() to go to another script, and then in that
script do some fancy notation, but the last line had to be text() to
start the internal scripting exploit or else it wouldn't work. Ved came
along and automatically did the notation for you.

A quick summary of why that worked: tokenize() is called on every line
even if it's not a valid simplified line, due to sheer dumb luck
customiftrinkets() comes before text() in the frame order and is in a
different "if"-"else if" tree than the other commands, so the game
continues processing the 'words' vector after customiftrinkets() is ran
and ends up processing the text() on the last line of the script.

Once upon a time this used to be useful, but given that in VCE you can
simply use internal commands in simplified, this whole ordeal is no
longer necessary. And the bug that enabled this whole thing now simply
gets in the way of actual scripting.

Fixing this is simple enough, just replace the 'position--' with a
'return'. The 'position--' is only there to undo the 'position++' at the
end of the loop because the function doesn't stop there, but if it does
stop there then the 'position--' is no longer necessary and in fact is
actually bad.

To maximize backwards compatibility with levels made before VCE, the
vanilla commands are not touched at all. So the only commands this is
fixed for are:
 - ifflipmode
 - iftrial
 - ifkey
 - ifvar (if() is an alias)
 - ifnotflag
 - ifcrewmates
 - ifcoins
 - ifcrewmatesless
 - ifcoinsless
 - ifrand
 - ifvce
 - ifmod
 - load
 - jump

On a side note, I removed the space-indentation and replaced it with
tabs for ifflipmode and ifmod.