Remove one level of indentation from ifmod()
Seems like Steljpo hasn't really internalized what the "and" conjunction
really is. 'if (cond1 && cond2)' is really just a shorthand for 'if
(cond1) { if (cond2) { stuff; } }', in most cases.

It isn't strictly a shorthand, sometimes you DO want the
if-cond1-then-if-cond2 instead. It's probably best to not collapse
conditionals onto the 'if (words[0] == thing)' line, for instance. But
INSIDE that 'if (words[0] == "ifmod")' block, I find that the extra
level of indentation is unnecessary.