Implement global exception handlers
There is an exception handler in main that will exit the program after
graphically displaying an error. In addition, there is now a handler in
scriptclass that will display an error along with the line and script
before exiting the level (in the same way gamestate 1013 would, except
for not touching stars).