Add fake levels/special/stdin.vvvvvv file
This can be used when playtesting as `VVVVVV-CE -p special/stdin` to
play a level piped to stdin. I loaded it into a buffer the "C++ way,"
which is much less verbose than the C way. However, the C way is
*arguably* easier to read (it probably makes exactly as much sense, but
it uses much simpler syntax). I believe this is C++98, so that shouldn't
be a concern.