Remove "list is updated with each major release"
This would require FEATURES.md be out-of-sync with the in-game
changelog, due to the fact that we are doing prereleases.

Maybe we can sort out whether or not this should be the case, but right
now it'll be a bigger headache if FEATURES.md and the in-game changelog
are not in sync, so I'm removing this sentence.