Clarify "all limits have been removed" in changelog
Nicholas brought this up and it is probably the pettiest complaint I've
seen.

But it's a fair point, because although we'd like to think we've
magically got all limits removed due to Leo's magical growing_vector, in
reality there are some limits that are not based on vector length, so we
can't really say that all limits are removed.

Instead I'll just list the big ones that VVVVVV levels usually run up
against, especially the 500 scripts and 500 script lines limit. I listed
"3000 edentities" as "3000 placeable entities" just because people may
not be familiar with edentities. I would've included the 500 blocks
limit but then I'd have to clarify what a block is and I don't have
enough space for that.