Add functionality for extended tiles*.png graphics
This commit adds the functionality of using files like tiles4.png,
tiles5.png, tiles6.png, etc. to allow for unlimited custom tilesets
within a level.

The approach is that we only change how the tiles are rendered, not the
types of tiles that they are. Instead, we let the user choose the tile
types using the existing tiles.png/tiles2.png/tiles3.png files,
depending on which tileset they're using in their room, and simply swap
out the graphics with another tilesheet instead.

To do this, I've added the room property 'customtileset'. Ved will have
to update accordingly. I considered making the custom tilesets only
available through the use of a script command, but that would've been
harder to visualize in the editor. We can always add said script command
later, to let level creators do some insane tileset swapping or
animations or something, but not right now.

I've added the following functions:
 - FILESYSTEM_getGraphicsDirFileNames()
 - Graphics::makecustomtilearray()
 - editorclass::getcustomtiles()

I store the tiles in Graphics::customtiles and
GraphicsResources::customtiles.