Fix fully-remove-1x1-quicksand code happening in vanilla levels
This commit fixes the fact that the fully-remove-1x1-quicksand code was
happening in vanilla levels. It was copied over from the code for
disappearing platforms for the same reason they have it - because if you
didn't fully remove a platform in the middle of being destroyed, you
could break the game when the entity respawned.

However in vanilla, 1x1 quicksand don't respawn anyway, and so it's fine
to let them continue their animation of being destroyed instead of
moving them right along to already being destroyed when the player dies.

This change also fixes an infinite loop that could occur if the player
died while a 1x1 quicksand is disappearing, due to the fact that
entityclass::updateentities() would be repeatedly being called on a
deactivated object, thus doing nothing, while inside a while-loop that
expected that function to do something.