Add optional args to setcheckpoint() to set checkpoint anywhere
You can now use setcheckpoint() like setcheckpoint(1,1,100,100,1) to set
the checkpoint to 100,100 in (1,1) (0-indexed), flipped. This also
supports relative tilde syntax relative to the current checkpoint, not
relative to the current position of the player. If an argument is
blanked out, it will default to the current position of the player.