Use main state for creating entities if alt state is invalid
If the current alt state is invalid when loading a room, i.e. the room
doesn't have an alt state with that state number, then it would default
the main state tiles, but it wouldn't create any entities at all in the
room, because creating entities still checked if the state of the entity
was in the invalid state, and since it's an invalid state it would
always fail to create the entity.

This commit makes it so that if the current alt state is invalid when
creating the entity, it will fall back to using the main state (state
0), and check entities against that state instead.