Allow using custom tilesheets in backgrounds
The game uses tiles2.png for the horizontal and vertical warp
backgrounds, and tiles3.png for the tower background (duh). Previously
they would always default to the default tilesheet, but this commit
fixes that. But the Tower Hallway background was unaffected and working
properly.

Also had to set graphics.backgrounddrawn to false when changing the
tilesheet in the editor, else the background wouldn't be updated
instantly and would slowly scroll away when the new textures came in.