Remove ifvar() from don't-parse-textboxlike-lines code
It hasn't been in there ever since the switch to the new ifvar() syntax.
So (1) the argument the code looks at would be wrong and (2) it also
just doesn't use the next line at all.

Come to think of it actually, why did it ever use the next line in the
first place?