Don't add -Werror=implicit-fallthrough on Android at all
It'll trip up in SDL stuff even though we only want it to be enabled for
only VVVVVV-CE itself. Since we can't do this (the 'VVVVVV-CE' project
doesn't exist for Android in the CMake, only 'main'), I'll just disable
it entirely. -Werror=implicit-fallthrough is only needed for development
purposes anyway, and you should be developing on a PC anyways. I've
never heard of anyone having a cursed setup where they develop without
using a PC.