Work around gcc bug
Lua requires that the user specifically opt in to C89, and uses the
existence of LLONG_MAX to determine that the compiler is C99. This check
is skipped over for Windows and POSIX. On the version of gcc used for
the Switch, LLONG_MAX is defined based on the __ISO_C_VISIBLE macro.
That macro is defined in sys/features.h. However, the gcc developers
forgot to include that header...