Refactor custom scripts to not be stored in one giant vector
This is the same thing I did upstream. Scripts are stored in a `Script`
struct, with a `name` and `contents`. `script.customscript` is gone,
instead there's a `script.customscripts` that is an
`std::vector<Script>`. This results in having less boilerplate to deal
with the giant pool of script lines, the only boilerplate code existing
in saving and loading the level. This makes having specifically-marked
Lua scripts much easier to deal with.

Just like upstream, I tested it by first saving Dimension Open in a
current build of VCE, and then re-saving it on this commit, and diffing
the two. There is no difference, which means I did everything right.