Distinguish Lua scripts by attribute instead of "#lua" contents prefix
Prefixes are a bit ew, so it's fancier to use an attribute to mark it
instead. After all, we have a `Script` struct, don't we?

This also fixes it so that previous Lua scripting code now works
properly with the script data structure refactor I did on master (and
upstream).