Refactor layer names to not be stringly-typed
String-typing is bad. One misspell and you could manifest a bug that
only reveals itself at runtime! If you misspell an enum however, you get
a compile error, which is a bug that reveals itself as early as
possible.

But how do we convert from strings (drawimage()/drawimagepersist()
argument) to the enum, without the dreaded practice of copy-and-pasting?
Why, we use X macros, of course! LAYERNAMES is one such X macro, and I
use it to (1) build the actual enum itself, and (2) build the
std::unordered_map lookup table to convert the string names to actual
enums. No copy-paste needed!