Remove tests
According to leo60228 himself (who added the tests in the first place),
there have been 20 times the tests have failed due to issues with the
tests, and only 1 time the tests have failed due to a legitimate bug in
VCE.

So it's better to just remove it. That way all the kludge VCE has to add
in order for the tests to work is gone. In particular, --headless,
--quiet, and the debug commands are now gone, so the codebase has been
cleaned up a bit. --quiet didn't work ALL the time anyway, not all print
calls have been wrapped with the --quiet check.